Database repairs and restructure to db rules.

+++ src/s2s_transport.c
@@ -521,7 +521,7 @@ s2s_send_json (s2s_conn_t *c, json_t *obj, int timeout_ms)
   size_t len = strlen (payload);
 
 
-  if (len > g_cfg.s2s.frame_size_limit)
+  if (len > (size_t) g_cfg.s2s.frame_size_limit)
     {
       LOGE ("s2s_send_json: frame too large (%zu > %d)",
 	    len, g_cfg.s2s.frame_size_limit);
@@ -569,7 +569,7 @@ s2s_recv_json (s2s_conn_t *c, json_t **out, int timeout_ms)
   uint32_t len = ntohl (be);
 
 
-  if (len == 0 || len > g_cfg.s2s.frame_size_limit)
+  if (len == 0 || len > (uint32_t) g_cfg.s2s.frame_size_limit)
     {
       LOGE ("s2s_recv_json: invalid/large frame length (%u, limit=%d)",
 	    len, g_cfg.s2s.frame_size_limit);
